requirements.txt
Django
djongo
djangorestframework
django-cors-headers